chore(node): centralize block creation loop
The compute worker had a fully separate block proposer and loop for generating blocks which made it hard to reason through times where both tax and compute were operating. This merges the two into a single loop that can be found in the entry point of the file. Additionally, this takes on it’s own responsibility for scheduling the essential handles for the threads

fix(node): wait for 2/3 of a tick before starting compute
We were beginning compute right away, which would cause it to be too short a period to include notebooks on time.