fix(node): default block votes
Notebooks are required to submit a default vote if and only if there are no other votes in a notebook. The notebook has no power, so will be beaten by any legiimtate vote. It must come from the account that is the operator of the notary (eg, cannot be manipulated). It also has no rewards. If a default block vote is used, it will give rewards for the compute portion, but the 1.25 argons will not be given to anyone.