chore(runtime): convert to token gateway
We’ve migrate to the token gateway pallet created by hyperbridge. In order to use the https://app.hyperbridge.network/ interface, we needed to use the stock pallet (vs our custom implementation). The primary impetus to using this is to enable state rollbacks, which are not implemented by default. It is also challenging to use the ethereum interface without some kind of interface, so we will want something, and a ux managed by a bridge provider is ideal.

The only features removed by using this approach are:
- No protection against attempting to register paths to send argons to testnet evms in prod and prod evms from testnet. However, this should be blocked at the hyperbridge level
- No “pause” bridge function is in place for this version. We can still pause teleports initiated on the argon side using the TxPause pallet