feat(notary): archive hosts
This PR adds the idea of archive hosts for a notary to host notebooks and headers for download, relieving the notary server from needing to handle traffic for those requests. Additionally, this allows a notary to reboot while keeping the notebooks available to the mainchain.

Secondarily, there is a broader notebook archive idea that will be produced by indviduals or groups in the network that can serve as a way to read old notebooks from archived notaries. Or a miner could choose to download all notebooks for themself to sync up.