fix(ticks): max 5 blocks per tick
Check before creating block if it will be able to submit a tick to the “blocksattick” in the next block. If not, it can stall the blockchain. As an addition, there’s now a 5th block allowed per tick, but it must have notebooks. This is to prevent a fast miner from stopping all notebooks from being included.