fix(seal_spec): trim to 80th pctl of block times
An issue was introduced from the requirement that the last block at a tick include notebooks where the difficulty adjustment is now measuring the time it takes to wait for the next block tick, which does not actually reflect the amount of block time.

This fix trims off the top 80th percentile of block times and computes difficulty from the rest. This will have the impact of increasing the time to equalize the compute difficulty for the fast compute nodes, but will slow down the time to reduce when the faster miners disappear. However, 80th percentile means this will mostly trap the block times for the first “4” in a tick and omit the one that has to wait for the next tick.

In addition, this change extends the “4” to 59, allowing 60 total blocks in a tick (or minute).