fix(oracle): handle us cpi resp annual in latest
The US CPI results contained an unexpected first oracle value for the year, which is an annual summary of last year as the “Latest cpi”. This data hadn’t been seen, so broke the api we were expecting. This changes us to use the regular data series (not the latest flag). It’s a little more data, but provides the dataset we’re looking for.