fix(node): grandpa can’t prove finality
The grandpa finality gadget will not yield proofs of finality until an authority rotation has occurred. This change forces an authority change - both when the new mining slots change, as we as via a migration. NOTE: authority changes in the same block do not seem to work, so these are scheduled for the next block.