fix(block_seal): don’t allow compute on vote
A compute block should not be allowed in the same tick as a vote block, or we would be allowing compute to control the duration of slots by maximizing blocks in that tick.

Compute can still capture blocks when vote miners are not capturing possible blocks, which would mean compute could impose up to a 60 minute time penalty if miners are not online to close blocks and no one else steps up to do so.