fix(notary): don’t fail audit for wrong tick
When a notebook comes in, it currently doesn’t have a good tracking for which block it can audit against. In longterm, we need to add a finalized notebook number to notebooks so we know what needs to be in the runtime to audit. Otherwise, we need to track a minimum tick needed