fix(node): don’t store bad sync for notary client
The node has a bug where the notary client will store bad sync state for a notebook if it has imported a block already containing a notebook at the time it goes to import the block. This appears to be some kind of race condition that presents when you first reboot a node and it is importing blocks at the same time it catches up notebooks.

The fundamental issue is that a notebook must be audited on blocks prior to being included in the state, so that’s what this fix does.