chore(notary_audit): don’t audit vote minimums
The notary should not be responsible for auditing blocks that are voted for. That should be the responsibility of the node, which should determine if it can build on a valid block and if the vote amounts were legitimate.

The only rule the notary shoud check is that the absolute minimum is followed. That way we can at least ensure votes can’t be faked out to try to change hashes.

The block seal process will still require a minimum vote