fix(mining_slot): miners losing lookup index
Mining slots had a bug where if miners re-entered the same slot with a different index, they could end up getting their lookup removed and breaking the ability to create blocks.

fix(vaults): bid pool allocation errors
The bid pool also had an error where it could end up not having access to the last 10% of the pool at time of allocation. The change is to call creating the new bid pool at the time of creating a slot. Effectively making the bid pool entrants bidders for the bid pool that starts with a slot rotation.