feat(mining-bonds): Mining Bonds
Mining bonds allow any account to add capital to earn a share of the mining bid pool. The vault specifies the split between themself and the contributors, and each contributor earns a share proportional the capital they lock up. The maximum capital is the activated securitization of the vault (eg, locked bitcoin * securitization ratio).

Vaults are simplified in this pr to a securitization amount and a securitzation ratio. Only up to 1/ratio funds can be locked as bitcoin, while the rest serve as fraud recovery funds.