feat(bitcoin_lock): ratcheting
Adds bitcoin ratcheting. Bitcoins can ratchet up price and capture mint value of difference, or ratchet down and burn existing funds to reset the lock price to the new low.

To enable vaults to do this, we had to implement overdue features for Vaults to:
a) re-lock released bitcoins, ensuring locks are held for a minimum of a year
b) count all released bitcoins that still hold locked funds as eligible for liquidity pools