fix(bitcoin_locks): truncation in redemption price
The redemption price calculation was doing a premature conversion to u128 which was causing truncation during fixed point math activities.

chore(bitcoin_locks): rename price provider fields
During the course of this fix, ti was also noticed that many of the variables in the provider incorrecly refer to the price index fields as being in cents, where they’re actually fixued128 of the usd amount, so the naming was updated.

fix(bitcoin_locks): The redemption_r_value was incorectly using argonot price.
The redemption value was using argonot, which caused calculations to be wrong.