feat(node): allow multiple miners to submit a vote
This change makes the system far more resilient. Instead of each vote allowing exactly one miner to submit, we now allow any miners in the system to submit a block for a vote. The system will prioritize the lowest xor distance. So the same miner as previously could win the block, but if they are unresponsive, the system will not default into compute blocks.

To avoid flooding the network, each node will only submit blocks after a delay according to their percentile of the xor order.

There are several major fixes uncovered by this change as well:
1. The sort order of forks was incorrectly keeping around the seal strength
2. Our “tie-breaker” did a direct equals comparison instead of a weight comparison
3. Our “best block beating algorithm” recursed the whole tree of blocks, resulting in a spidering out that created even more blocks. This was refined to only looking at the best block at a new notebook’s tick and seeing if we can create a better child block.
4. The sort order of forks meant that building compute on top of voting could stall because not vote could find a better seal strength than the “straggling” seal strength that was not properly cleared out.