critical: patch memory issues in random-x
During testing of the 1.3.0 release, a critical memory error in the proof of compute randomx library was found. Further analysis showed a few issues that needed to be addressed. This pulls in the latest branch on the argon randomx library (memory-safety) and clears up some intent of the library in this project so that it is not used in a cross-thread way.