feat(vaults): earnings must be collected
This feature introduces the idea of “collect” for vaults. Vaults must now collect revenue every 10 frames or forfeit it. This is to ensure vaults have the right incentive to cosign anything needed. To accomodate this feature, all revenue is tracked in the Vaults#PerFrameRevenueByVault field.

Also as part of this, bitcoin cosign requests now expire in 10 frames instead of 10 bitcoin days to ensure they are aligned.