fix(runtime): remove 1 second minimum time
The timestamp pallet was bumping out timestamps by a minimum of 1 second per block if it was shorter. This caused development mode of the blocks to start rejecting blocks. Instead of using this timestamp feature, we should rely on the other built-in processes to stop blocks from being accepted - such as only 1 vote block per tick, and 60 max blocks per tick that (capped in the runtime).

All blocks must continue to have a timestamp in future of the previous timestamp