feat(bitcoin): track locked liquidity value
Bitcoin locks need to track the lock price separately from the liqudity value. The lock price is a cpi adjusted value in microgons for a bitcoin lock, and this value must be used at release time to determine the original “true” value. This is separate from the liquidity value, which is simply the value at time of locking that will be minted to the user.