fix(node): hanging sync for notary
The notary sync process had two flaws that could cause it to hang during initial sync or put the sync into a bad state.

The verify headers process runs for the first imported block, but state has only downloaded for a block that is now old. In this case, the best block state is not available, but it has synced state to some point. At this point, new blocks that are verified should NOT unlock the lock for all processes like mining, but only an import workaround. And this import workaround needs to only use the parent hash as the best block, not the current best block.