fix: update distribution of miners chosen
The xor closest algorithm used to pick the most eligible miner to submit each block was not providing a clean distribution. This changes the process to account for how recently each miner has created a block and uses a hash instead of an xor to create a more uniform distribution across all miners (near 1-2% of a clean split).